Moon repository completed with variables. Integer changed for PlanetRepo.

+++ src/main/java/org/example/springframeworkassignment/repositories/PlanetRepository.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Function;
 
-public interface PlanetRepository extends JpaRepository<Planet, Long> {
+public interface PlanetRepository extends JpaRepository<Planet, Integer> {
 
     List<Planet> findPlanetByType(String planetType);
 
@@ -28,6 +28,5 @@ public interface PlanetRepository extends JpaRepository<Planet, Long> {
     @Transactional
     @Query("DELETE FROM Planet p where p.planetID = :planetID")
     public int deletePlanetById(@Param("planetID") int planetID);
-
 
 }
